services/reverse-search/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.3
aiohttp==3.9.1

# New additions for enhanced search
google-search-results==2.4.2  # SerpAPI
imagehash==4.3.1
newspaper3k==0.2.8
lxml==5.1.0
python-dateutil==2.8.2
